rename index to component
